genicam: implement Cachable support
The work was started long time ago, but never used, and mostly broken.
This is a complete rewrite of Cachable support.

Caching is disabled by default for now, as devices seem to not always
define correctly the Cachable property. A debug mode is available, that
will compare the cache content with the actual register value, and show
a message on the console if they differ.